[TASK] Migrate Redirect Url Validator into own class
The logic within EXT:felogin for validating proper redirect/referer
URLs is extracted into a new PHP class, which can now be called
separately.

This is useful for further refactorings within the Controller,
and it lays the foundation to use proper validation in various
places other than Extbase.

Resolves: #87660
Releases: master
Change-Id: I79aab5908bde869b7ee3cf730fbf61d658f2ee88
Reviewed-on: https://review.typo3.org/59648
Tested-by: TYPO3com <noreply@typo3.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>