[BUGFIX] Make slugs unique when copying pages
When a page is copied the slug for the copied page
needs to be unique for a site. To accomplish that,
the site for the record is fetched. Since this new
record does not have a uid yet, the parent page uid
must be used to determine the site.

Resolves: #87697
Releases: master, 9.5
Change-Id: Iac5fa7adc673c41662b652e932c23b90633e0d5d
Reviewed-on: https://review.typo3.org/c/59686
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Jürgen Venne <venne@schaffrath-digital.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Jürgen Venne <venne@schaffrath-digital.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>