[BUGFIX] Have stdWrap.lang use Site concept
When building TypoScript like this:

10 = TEXT
10.value = English
10.lang.dk = danish

This does not work currently as it relies on "config.language". This part
has been forgotten while the Site Handling was initially implemented into
the TYPO3 core.

The patch adapts the changes to make use of Site Handling if it is available,
but falls back to the TypoScript options (for PseudoSites).

Resolves: #87712
Releases: master, 9.5
Change-Id: I5e2b97b178e6ccafec1ff7556d41321ea67fb45a
Reviewed-on: https://review.typo3.org/59704
Tested-by: TYPO3com <noreply@typo3.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>