[BUGFIX] EXT:opendocs - Remove bogus initial document count
When the opendocs toolbar item gets initialized, the document count shows
a bogus value as deleted records are not removed from the calculation.

As the documents are loaded deferred since #84412, the correct value gets
in place automatically which renders the initial counting obsolete and is
removed with this patch.

Additonally, a wrong class hint in OpenDocumentController is fixed.

Resolves: #87725
Related: #84412
Releases: master, 9.5
Change-Id: I0684e0950d0f4daf11042d6c078b5598bfb3d4be
Reviewed-on: https://review.typo3.org/59708
Tested-by: TYPO3com <noreply@typo3.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Björn Jacob <bjoern.jacob@tritum.de>
Reviewed-by: Tymoteusz Motylewski <t.motylewski@gmail.com>
Tested-by: Tymoteusz Motylewski <t.motylewski@gmail.com>