[TASK] Use real ObjectManager stubs in unit tests
Prevent invocation of the constructor (which instantiates
an internal instance of the extbase container) and do not
mix usage of mocked and stubbed methods.

The mixture happened due to the usage of setMethods():
When specifing `setMethods(['get'])` all methods besides ->get()
are actually proxied to the real ObjectManager methods.

Releases: master, 9.5
Resolves: #87742
Change-Id: I817f8ee90d82f36b8117e873b5802e11f601384e
Reviewed-on: https://review.typo3.org/c/59757
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>