[TASK] Do not instantiate DocumentTemplate at every BE call
Currently, document template and thus, pagerenderer is initialized
with all functionality and loaded CSS for every BE call,
which is unnecessary for simple JSON-based AJAX calls.

In order to avoid this, PageRenderer is now instantiated only
when needed - when ModuleTemplate or DocumentTemplate
are initialized, resulting in higher performance for AJAX calls
but also removing a necessary means for having DocumentTemplate
set up at all BE requests, reducing the need for DocTemplate
even more, so it can slowly be migrated to ModuleTemplate
in the future.

Resolves: #87766
Releases: master
Change-Id: Ic1549be7f4d62164bf99cb496fad4b96dee2bc7d
Reviewed-on: https://review.typo3.org/c/59781
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>