[BUGFIX] Add extension suggestions to EXT:form
EXT:form adds a ContextMenu provider which depends on classes of
EXT:filelist. As EXT:filelist is optional in Composer mode, add the
extension as a suggestion and register the ContextMenu provider only in
case the extension is loaded.

Similar guards are added for hooks of EXT:filelist and EXT:impexp.

Resolves: #87785
Releases: master, 9.5
Change-Id: I254137bc85c87c3d2b6571e32676e7a579f342a9
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/59807
Tested-by: Björn Jacob <bjoern.jacob@tritum.de>
Tested-by: Susanne Moog <look@susi.dev>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Ralf Zimmermann <ralf.zimmermann@tritum.de>
Reviewed-by: Björn Jacob <bjoern.jacob@tritum.de>
Reviewed-by: Susanne Moog <look@susi.dev>
Reviewed-by: Ralf Zimmermann <ralf.zimmermann@tritum.de>