[BUGFIX] Prevent exception in file list
When creating a file with a disallowed file
extension an exception was thrown as on creation
the redirect to edit interface on a non-existing
file was called.

Though the better fix might be to prevent the
request for an invalid file to be sent at all
this fix ensures a working file list module in
a more "surgical" way to allow secure backporting.

Resolves: #87527
Releases: master, 9.5, 8.7
Change-Id: I35a054c05b37c09acab83a7aa9eca89cf9ebf6b9
Reviewed-on: https://review.typo3.org/c/59926
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>