[BUGFIX] Ensure Lowercase folders on non-case-sensitive file system
When creating files in a storage configured as not case-sensitive
the file name gets lowercased and TYPO3 works with the lowercased
version everywhere. When creating folders the lowercasing was missing,
resulting in an exception as the expectation still was to have only
lowercased files and folders on non-case-sensitive systems.

Lowercasing for folder creation has been implemented.

Resolves: #85583
Releases: master, 9.5, 8.7
Change-Id: Icd12ea602ceda9de8bc1a44c58f8ebeaef9a8008
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/59931
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Susanne Moog <look@susi.dev>
Reviewed-by: Stephan Großberndt <stephan.grossberndt@typo3.org>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Susanne Moog <look@susi.dev>