[BUGFIX] Log invalid hash generation in FAL Indexer
The FAL Indexer uses the driver to get a hash for a file. In certain
circumstances (and with 3rd party drivers) the hash generated might be
null or invalid. In that case the ResourceStorage now throws an
exception to alert the developer to the wrong return value. At the same
time the Indexer catches and logs the exception to not interrupt the
whole indexing flow because of it.

For more information, please take a look at the ticket.

Resolves: #66397
Releases: master
Change-Id: I0f97cd23127e96b6d94161a9296750cbe3421075
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/62230
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>