[BUGFIX] Include Records with "All Languages" in default language
Records with "All Languages" (sys_language_uid=-1)
aren't shown in TYPO3 v9 anymore of the default
language.

This happened before in TYPO3 v8 and below,
however setting config.sys_language_overlay = 1
triggered the if() condition to jump into the
selection of fetching 0 and -1 records.

As this setting is removed, the check only
applies when overlays are activated, which
is not necessary in language=0.

Resolves: #86972
Resolves: #87749
Resolves: #87093
Releases: master, 9.5
Change-Id: I9308167aede071125df02b848a26283e3f77b759
Reviewed-on: https://review.typo3.org/c/60192
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>