[BUGFIX] Deliver detected file type for IM identify
Although not used in TYPO3 Core directly, the identify
command by ImageMagick now delivers the file extension
from IM plus the file type identified by IM as well.

This saves some regexp magic, and a foreach() loop,
as the "-format" output delivers proper IM results.

As this functionality is not actually fixing a core
bug, but rather a small improvement and correction
in the identified usage, the change is targeted to master-only.

Resolves: #23348
Releases: master
Change-Id: Ia380ee34dbbef7f9f763ecc088ee0c9b83167ce6
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/60311
Tested-by: Sebastian Rosskopf <sebastian.rosskopf@gmx.de>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Sebastian Rosskopf <sebastian.rosskopf@gmx.de>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>