[BUGFIX] Fix serialization of loggers
Only dealing with PHP file resource in writers is not
enough to properly initialize a logger on wakeup.

A logger has to log with a new request IDs and possibly completely new writers
and processors due to configuration changes in the meantime.

Therefore the __sleep and __wakeup methods are removed in the FileWriter
and new methods are added to the logger instance to make sure a
wakeup will have a correctly configured logger instance.

Resolves: #86941
Resolves: #87261
Releases: master, 9.5
Change-Id: Ia36e251404eae2bdf0dfdf52ace7dcf1815ec456
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/60472
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Susanne Moog <look@susi.dev>
Reviewed-by: Susanne Moog <look@susi.dev>