Fix scroll behaviour of "Uncaught Exceptions" in backend
In the TYPO3 Backend of version 8, the iframe page (id "typo3-contentIframe") is not scrollable, when an "uncaught exception" is thrown and its content exceeds the page. The simplest way to fix this behavior seems to be by adding 3 CSS rules to the very first div element after the body tag in the iframe page (right, max-height and overflow; "right" is needed to enable horizontal scrolling).

This solution is not perfect, when it comes to horizontal scrolling - but better than the current situation.