[TASK] Fix scroll behaviour of "Uncaught Exceptions" in backend
In the TYPO3 Backend of version 8, the iframe page (id
"typo3-contentIframe") is not scrollable, when an "uncaught exception" is
thrown and its content exceeds the page. The simplest way to fix this
behavior seems to be by adding 3 CSS rules to the very first div element
after the body tag in the iframe page (right, max-height and overflow;
"right" is needed to enable horizontal scrolling).

This solution is not perfect, when it comes to horizontal scrolling - but
better than the current situation.

Releases: 8.7
Resolves: #88096
Change-Id: I4ace93fc06fab48768944a5ef58958892ab51897
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/60410
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Jonas Eberle
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Jonas Eberle
Reviewed-by: Simon Gilli <typo3@gilbertsoft.org>
Reviewed-by: Benni Mack <benni@typo3.org>