[FEATURE] Add custom generator options via hook
In order to allow custom functionality like "Skip pages of type XYZ", a new
hook is added to the SlugHelper which can be added via

$GLOBALS['TCA']['pages']['columns']['slug']['config']['generatorOptions']['postModifiers']

The userFunc needs to return the new / modified $slug which is then put through the
sanitize() method.

TYPO3 Core does not provide custom functionality so this hook allows
to build everything a custom installation or extension wants to hook in.

Resolves: #88198
Releases: master, 9.5
Change-Id: Iad548da69ba001c696afd7d7f834d4950318c029
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/60628
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Susanne Moog <look@susi.dev>
Reviewed-by: Susanne Moog <look@susi.dev>