[!!!][FEATURE] Introduce migration wizard for changed flexform fields
Rename ext:felogin's flexform fields. Add prefix `settings.` to all of
them. This is for the later extbase change to easily access the flexform
values via $this->settings in the extbase controllers.

An update wizard is provided to migrate all existing felogin plugins
which are stored in the database.

Resolves: #88129
Releases: master
Change-Id: I76158d8c73d5510fc537fcd5d5bbb376b156c1e7
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/60450
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>