[!!!][TASK] Remove legacy "pageNotFound_handling" options
With the introduction of site handling, the old "pageNotFound_handling"
mechanisms can be removed. If a site configuration hasn't configured
a error handler, a 404/403/500 response with the default error
renderer is now returned. Making PageNotFoundExceptions rendering
the frontend very seldom.

The following options are removed:
- $TYPO3_CONF_VARS[FE][pageNotFound_handling]
- $TYPO3_CONF_VARS[FE][pageNotFound_handling_statheader]
- $TYPO3_CONF_VARS[FE][pageNotFound_handling_accessdeniedheader]
- $TYPO3_CONF_VARS[FE][pageUnavailable_handling]
- $TYPO3_CONF_VARS[FE][pageUnavailable_handling_statheader]

Tests have been adopted to check against status code / message
rather than on the exceptions (not) thrown (anymore).

In addition, a JsonResponse is now returned if JSON
was requested, instead of the regular HTML content.

Resolves: #88376
Releases: master
Change-Id: I6be57fadaa0ddf83039fbdcc9186ef93d5a51158
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/60453
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Benni Mack <benni@typo3.org>