[!!!][TASK] Remove database field "t3ver_id"
The database field "t3ver_id" which is managed by TYPO3 to store the
number of the version for a specific version is not in use anymore,
as this was related to the non-workspace managed versioning concept.

This linear versioning concept was removed, so the database field is
not needed anymore.

It is recommended for all extensions with TYPO3 v10.0+ compat only
to remove this field.

Resolves: #88143
Releases: master
Change-Id: I283e8e2ba2d755dcc8b2054e2439f9fef5462511
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/60459
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Achim Fritz <af@achimfritz.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>