[BUGFIX] Allow to check/uncheck all recipients in Workspace stage dialog
The stage dialog shows various users relevent in a workspace that
should be notified. However, it is hard for editors (UX wise) to
uncheck / check everybody when a single element should be pushed
into the next stage.

The patch adds two buttons to toggle the checkboxes (check/uncheck all).

Resolves: #59228
Releases: master, 9.5
Change-Id: I2ebbcb7ec2d46b50880f93aefd243ef791300a2e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/60496
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>