[BUGFIX] Sanitize undefined TCA columns required for data integrity
TCA's 'ctrl' section allows to define several database columns that
shall be used to store according integrity information, such as the
current language or pointers to ancestors used during localization.

In case those names are not defined properly in TCA's 'columns'
section, several commands (like copy of localize) are executed,
but without actually maintaining these values in the database.

In order to ensure integrity, missing columns that are defined in
the 'ctrl' section but missing in the 'columns' section are applied
with the TCA type 'passthrough'. This applies to 'ctrl' properties

* origUid
* languageField
* translationSource
* transOrigPointerField

Resolves: #88057
Releases: master, 9.5
Change-Id: I39a28dc2e1eddafe6363b7dd633fd84968fc620f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/60396
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>