[BUGFIX] Close file resource in FileLockStrategy if acquire failed
When acquiring a lock via the FileLockStrategy it is very important
to close the underlying file handle again, if the acquire action failed.

The same instance of the FileLockStrategy might be reused later
(e.g. in TSFE) again, whereas the (still open) file could have been
deleted meanwhile. The file handle would be invalid and another try
to open the same file again - with fopen() - may fail.
The behaviour of PHP's fopen() is different depending on the actual
file storage. Currently things fail on host-mounted drives in
containers in Docker for Windows.

Resolves: #88197
Releases: master, 9.5, 8.7
Change-Id: If802c670b617119d28aca09fcd5acef95f0ae678
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/60562
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Susanne Moog <look@susi.dev>
Reviewed-by: Susanne Moog <look@susi.dev>