[BUGFIX] Catch form errors and display speaking error message
When having a broken form configuration, the formEditor is not able
to load because of JavaScript errors. These errors are not fixable for
an editor as they most likely happen because of admins/integrators
changing and breaking the form. Instead of displaying a loading
indicator without a hint on the error that happened, a top level
catch displaying a flash message has now been implemented.

Resolves: #84595
Releases: master, 9.5
Change-Id: I44cf49d8e344629bc1a3250eade6267c7debdf00
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/60598
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Mathias Brodala <mbrodala@pagemachine.de>
Tested-by: Ralf Zimmermann <ralf.zimmermann@tritum.de>
Reviewed-by: Mathias Brodala <mbrodala@pagemachine.de>
Reviewed-by: Ralf Zimmermann <ralf.zimmermann@tritum.de>