[BUGFIX] Make sure a twitter:card metatag is rendered
To get a valid Twitter card, you need to specify the type of card
by setting the twitter:card metatag. The value of this metatag is
now set to summary by default. If this value need to be changed, you
can use the MetaTag API and set the twitter:card property with the
replace option enabled.

Resolves: #88258
Releases: master, 9.5
Change-Id: Iaeaafb2408102c2844623155611cffffa8919d43
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/60657
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Frank Naegler <frank.naegler@typo3.org>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>