[BUGFIX] Consider hidden records in persisted mapper aspects
PersistedAliasMapper and PersistedPatternMapper did not take
hidden records into account.

The Context object used when resolving and generating routes is
forwarded to corresponding mapper aspects which are applying this
information in their FrontendRestrictionContainer.

Only the visibility aspect is considered for HiddenRestriction,
FrontendGroupRestriction, StartTimeRestriction, EndTimeRestriction
are not taking the Context object into account, yet.

Resolves: #88307
Releases: master, 9.5
Change-Id: Iebddb32ba534002a19f458cb69947e4a2d44e990
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63642
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>