[!!!][TASK] Remove config.defaultGetVars
The TypoScript option config.defaultGetVars is not
evaluated anymore, as this conflicts with TypoScript
logic itself:

After TypoScript parsing and condition evaluation in
a Frontend Request, config.defaultGetVars is set,
this would effectively mean that TypoScript
parsing should be re-done in order to meet new
possible matches in Conditions.

As this logic is flawed and should now easily be
handled via a custom middleware (previously this was
helpful for setting e.g. L=0 for default pages
with RealURL), it is now removed from Core.

Resolves: #88419
Releases: master
Change-Id: Ic286c96f71953062d7649d71abb72f8f0f20f1a9
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/60799
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Susanne Moog <look@susi.dev>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Susanne Moog <look@susi.dev>