[BUGFIX] File can be deleted even if there are references to it
TYPO3 prevents files from being deleted by the user when there
are references to it in the sys_file_reference table. However
if the file is referenced without using this table, the file is
still deleted without any warnings. An example is a link to the
file in a bodytext of a text content element. Thus it is possible
to accidentally delete a file, which is referenced on the site
in multiple places and create broken links.

This change makes sure that such direct references (without using
sys_file_reference) also prevent the file from being deleted.

Resolves: #88462
Resolves: #77866
Releases: master, 9.5, 8.7
Change-Id: Ib8c45863d75ae430c801173658cf058f850ca982
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/60846
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Steffen Frese <steffenf14@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Jörg Bösche <typo3@joergboesche.de>
Reviewed-by: Julian Geils <j_geils@web.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>