[BUGFIX] Discard published drafts of workspaces when not swapping
When a workspace record is published, the "old", then unused record
has the values:
- workspace-id: 0
- online-id: > 0
- pid: -1

This way, it is not possible to actually find out from which workspace
something was published. This only happens when
not swapping (exchanging data) but doing real "publishing"
from a workspace.

What happens is that the database is filled with obsolete "previously
online" records that have no value anymore.

These records are now deleted, thus, the database is kept
clean. Previously only a CLI command allowed to remove
them, however, they are not needed at all, so the CLI command
is for existing projects that use workspaces in non-swapping mode.

Resolves: #88497
Releases: master
Change-Id: I9393c801de5b62484efc7ef82e389dfff6715a78
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/60882
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>