[BUGFIX] Ensure FormEngine IRRE handles relations in workspace
FormEngine does not show and handle IRRE records when they are
deleted (connected records from the live workspace are shown).

When removal and refactoring of code in this ares was done for
v9, "too much" code got removed. Which is why workspace overlays
are not handled properly anymore.

In addition, the ElementInformationController now correctly
resolves a workspace record when finding the pageInfo and access.

Resolves: #88502
Related: #82634
Releases: master, 9.5
Change-Id: Ib2d70ffa92c78752626b43ec5012d9dba036e7b8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/60890
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Achim Fritz <af@achimfritz.de>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Achim Fritz <af@achimfritz.de>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>