[BUGFIX] Always provide default language
The TranslationConfigurationProvider should always provide the default
language, even if the backend user doesn't have permissions configured.

This is due to the fact that some actions always require the presence
of the default language. This is a regression introduced with sites.

This patch always adds the default languages, as it was with pseudo
sites in place.

Resolves: #88504
Releases: master, 9.5
Change-Id: I2a8be4d399f7c24f8e2f8ec53374d9f419e92b71
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/61252
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Richard Haeser <richard@maxserv.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>