[TASK] Remove dependencies of \TYPO3\CMS\Extbase\Mvc\Web\Request
According to PSR-7 request objects should be immutable.
That can only be achieved if the state of the request
object is defined during construction. Therefore, the
web request object of Extbase mustn't have any service
dependencies that alter the state of the object during
runtime.

Although there are still setters that can alter the
state of request objects, the more crucial issue with
the object state were the injected services.

Releases: master
Resolves: #88697
Change-Id: Iedf6fc3d33a1e5b89bce20019cbfc484b1466cf8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/61235
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Jörg Bösche <typo3@joergboesche.de>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>