[BUGFIX] Prevent separating HTML inline elements
The softhyphen button in RTE CKEditor separates HTML inline elements.
If the softhyphen is inserted inside a link, it results in two
individual links with the &shy; entity between.

This can already be examined in the source view of the RTE.

Furthermore, a &shy; entity between two HTML tags will prevent the
browser to insert a hyphen at the position of the linebreak.

Setting the optional mode 'text' to the insertHtml event will prevent
this behaviour, as the inserted content will then inherit the styles
which are applied in the selected position.

Resolves: #88727
Releases: master, 9.5, 8.7
Change-Id: I4a398b02aba306016da12f290dab958b2e9ccf2c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63306
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Susanne Moog <look@susi.dev>
Reviewed-by: Susanne Moog <look@susi.dev>